requirements.txt
Django==4.1.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
pymongo==4.2.0
sqlparse==0.2.4
python-decouple==3.6
python-dotenv==0.21.0